refactor project

+++ src/main/java/com/github/srg13/socialnetwork/repository/UserRepository.java
@@ -1,9 +1,9 @@
 package com.github.srg13.socialnetwork.repository;
 
-import com.github.srg13.socialnetwork.domain.User;
+import com.github.srg13.socialnetwork.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByUsername(String username);
